demo changes

+++ Packs/Core/Integrations/CortexCoreIR/CortexCoreIR.py
@@ -1892,13 +1892,15 @@ def remove_blocklist_files_command(client: Client, args: Dict) -> CommandResults
                                         detailed_response=detailed_response)
     if isinstance(res, dict) and res.get('err_extra') != "All hashes have already been added to the allow or block list":
         raise ValueError(res)
-    markdown_data = [{'fileHash': file_hash} for file_hash in hash_list]
+    markdown_data = [{'removedFileHash': file_hash} for file_hash in hash_list]
 
     return CommandResults(
         readable_output=tableToMarkdown('Blocklist Files Removed',
                                         markdown_data,
-                                        headers=['fileHash'],
+                                        headers=['removedFileHash'],
                                         headerTransform=pascalToSpace),
+        outputs_prefix=f'{INTEGRATION_CONTEXT_BRAND}.blocklist',
+        outputs=markdown_data,
         raw_response=res
     )
 
@@ -1929,12 +1931,14 @@ def remove_allowlist_files_command(client, args):
                                         comment=comment,
                                         incident_id=incident_id,
                                         detailed_response=detailed_response)
-    markdown_data = [{'fileHash': file_hash} for file_hash in hash_list]
+    markdown_data = [{'removedFileHash': file_hash} for file_hash in hash_list]
     return CommandResults(
         readable_output=tableToMarkdown('Allowlist Files Removed',
                                         markdown_data,
-                                        headers=['fileHash'],
+                                        headers=['removedFileHash'],
                                         headerTransform=pascalToSpace),
+        outputs_prefix=f'{INTEGRATION_CONTEXT_BRAND}.allowlist',
+        outputs=markdown_data,
         raw_response=res
     )
 
@@ -2642,8 +2646,12 @@ def run_script_kill_process_command(client: Client, args: Dict) -> List[CommandR
 
 def add_exclusion_command(client: Client, args: Dict) -> CommandResults:
     name = args.get('name')
-    status = args.get('status', "ENABLED")
+    if not name:
+        raise DemistoException("Didn't get name arg. This arg is required.")
     indicator = args.get('filterObject')
+    if not indicator:
+        raise DemistoException("Didn't get filterObject arg. This arg is required.")
+    status = args.get('status', "ENABLED")
     comment = args.get('comment')
 
     res = client.add_exclusion(name=name,
@@ -2660,11 +2668,14 @@ def add_exclusion_command(client: Client, args: Dict) -> CommandResults:
 
 def delete_exclusion_command(client: Client, args: Dict) -> CommandResults:
     alert_exclusion_id = arg_to_number(args.get('alert_exclusion_id'))
+    if not alert_exclusion_id:
+        raise DemistoException("Didn't get alert_exclusion_id arg. This arg is required.")
 
     res = client.delete_exclusion(alert_exclusion_id=alert_exclusion_id)
 
     return CommandResults(
         readable_output=f"Successfully deleted the following exclusion: {alert_exclusion_id}",
+        outputs={f'{INTEGRATION_CONTEXT_BRAND}.deletedExclusion.rule_id(val.rule_id == obj.rule_id)': res.get("rule_id")},
         raw_response=res
     )
